legal-doc-generator/requirements-prod.txt
# FastAPI and Server
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
pydantic>=2.6.0
pydantic-settings>=2.1.0
jinja2>=3.1.2
aiofiles>=0.8.0

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
# For MySQL (alternative on Hostinger)
pymysql>=1.1.0

# PDF Processing
python-docx>=0.8.11
PyPDF2>=3.0.0
reportlab>=4.0.4

# AI Services
google-generativeai>=0.8.4

# Security
python-jose>=3.3.0
passlib>=1.7.4
python-dotenv>=1.0.0

# Logging
loguru>=0.7.2 